refactor: 자동/수동 로또 합쳐주는 메서드 클래스로 분리
