+++ src/main/java/Lotto/domain/LottoConcat.java
@@ -1,0 +1,13 @@
+package Lotto.domain;
+
+import java.util.List;
+
+public class LottoConcat {
+    public static Lottos concatLottos(Lottos autoLottos, Lottos manualLottos) {
+        List<Lotto> allLotto = autoLottos.getLottos();
+        if (manualLottos != null) {
+            allLotto.addAll(manualLottos.getLottos());
+        }
+        return new Lottos(allLotto);
+    }
+}
